Correccion de panelHomeSecudario
