+++ eventManager/WinFormsApp1/Vistas/PanelHomeSecudario.resx
